# docs/openapi.yaml
openapi: 3.0.3
info:
  title: CleanCast API
  description: |
    CleanCast is a podcast API that converts YouTube channels and playlists into podcast feeds.
    It automatically removes sponsored segments using SponsorBlock data and provides audio streaming.

    ## Features
    - Convert YouTube channels and playlists to RSS feeds
    - Automatic sponsor segment removal
    - Episode recommendations and trending content
    - Playback analytics
    - Audio streaming with range request support
  version: 1.0.0
  contact:
    name: API Support
    url: https://github.com/ikoyhn/clean-cast
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8080
    description: Local development server
  - url: https://api.cleancast.example.com
    description: Production server
...
paths:
...
  /channel/{channelId}:
    get:
      summary: Get channel RSS feed
      description: |
        Returns an RSS feed for a YouTube channel with sponsor segments removed.
        Episodes are automatically filtered based on minimum duration settings.
      tags:
        - podcasts
      parameters:
        - $ref: '#/components/parameters/ChannelId'
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Date'
      responses:
        '200':
          description: RSS feed XML
          content:
            application/rss+xml:
              schema:
                type: string
                format: xml
        '400':
          description: Invalid request parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized - API key required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '429':
          description: Rate limit exceeded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
        - ApiKeyAuth: []
  /rss/{youtubePlaylistId}:
    get:
      summary: Get playlist RSS feed
      description: |
        Returns an RSS feed for a YouTube playlist with sponsor segments removed.
        All episodes in the playlist are included.
      tags:
        - podcasts
      parameters:
        - $ref: '#/components/parameters/PlaylistId'
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Date'
      responses:
        '200':
          description: RSS feed XML
          content:
            application/rss+xml:
              schema:
                type: string
                format: xml
        '400':
          description: Invalid request parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized - API key required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '429':
          description: Rate limit exceeded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
        - ApiKeyAuth: []
  /media/{youtubeVideoId}:
    get:
      summary: Stream episode audio
      description: |
        Streams the audio file for a specific episode.
        Supports HTTP range requests for seeking.
        Downloads and processes the video on first request.
        Automatically removes sponsor segments using SponsorBlock data.
      tags:
        - media
      parameters:
        - name: youtubeVideoId
          in: path
          required: true
          description: YouTube video ID (11 characters)
          schema:
            type: string
            pattern: '^[a-zA-Z0-9_-]{11}$'
            example: "dQw4w9WgXcQ"
        - name: Range
          in: header
          required: false
          description: HTTP range header for partial content requests
          schema:
            type: string
            example: "bytes=0-1023"
      responses:
        '200':
          description: Full audio file
          content:
            audio/mp4:
              schema:
                type: string
                format: binary
        '206':
          description: Partial content (range request)
          headers:
            Content-Range:
              schema:
                type: string
              description: Range of bytes being returned
          content:
            audio/mp4:
              schema:
                type: string
                format: binary
        '400':
          description: Invalid video ID
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized - API key required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Video not found or download failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '429':
          description: Rate limit exceeded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
        - ApiKeyAuth: []
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key for authentication
  parameters:
    ChannelId:
      name: channelId
      in: path
      required: true
      description: YouTube channel ID (typically starts with UC)
      schema:
        type: string
        pattern: '^UC[a-zA-Z0-9_-]{22}$'
        example: "UCXuqSBlHAE6Xw-yeJA0Tunw"
    PlaylistId:
      name: youtubePlaylistId
      in: path
      required: true
      description: YouTube playlist ID
      schema:
        type: string
        example: "PLrAXtmErZgOeiKm4sgNOknGvNjby9efdf"
    Limit:
      name: limit
      in: query
      required: false
      description: Maximum number of episodes to include in the feed
      schema:
        type: integer
        minimum: 1
        maximum: 100
        example: 10
    Date:
      name: date
      in: query
      required: false
      description: Filter episodes published after this date (format MM-DD-YYYY)
      schema:
        type: string
        format: date
        pattern: '^\d{2}-\d{2}-\d{4}$'
        example: "01-01-2024"
  schemas:
...
    ErrorResponse:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
          description: Error code for programmatic error handling
          example: "INVALID_PARAM"
        message:
          type: string
          description: Human-readable error message
          example: "Invalid parameter provided"
        details:
          type: object
          additionalProperties: true
          description: Additional error details
          example:
            parameter: "channelId"
            value: "invalid-id"
        request_id:
          type: string
          description: Request ID for debugging
          example: "req-123456789"
...